Dynamic setting of params in speckle filter
Done previously by berend-kupers, but removed during code review because
we thought it needed a separate node inside the filter. Here we use the
params_interface_ member variable of the filter base class instead of
creating a new node.

The callback needs to be post_set_params and not on_set_params because
the changes haven't been finalized yet when on_set_params is called.